+++ guillotina/tests/dbusers/test_manage_groups.py
@@ -79,7 +79,7 @@ settings_with_catalog["applications"].append("guillotina.contrib.catalog.pg")
 settings_with_catalog.setdefault("load_utilities", {})  # type: ignore
 settings_with_catalog["load_utilities"]["catalog"] = {  # type: ignore
     "provides": "guillotina.interfaces.ICatalogUtility",
-    "factory": "guillotina.contrib.catalog.pg.PGSearchUtility",
+    "factory": "guillotina.contrib.catalog.pg.utility.PGSearchUtility",
 }